rename: 오타 수정(#146)
