+++ src/features/player/components/NotFoundSearch.tsx
@@ -1,4 +1,4 @@
-function NotFonudSearch() {
+function NotFoundSearch() {
   return (
     <p className="h-screen text-wiz-white font-bold text-center pt-4 text-xs md:text-base lg:text-xl">
       검색 결과가 없습니다.
@@ -6,4 +6,4 @@ function NotFonudSearch() {
   );
 }
 
-export default NotFonudSearch;
+export default NotFoundSearch;
